Fix CI: widen _scenario test helper to accept float counts
The new degenerate-count tests pass float values (inf/nan/1e308) to the
_scenario helper, whose **items: int signature mypy (CI, clean cache)
rejected. Widen to float — int literals stay compatible.

+++ tests/test_l2_silo_modules.py
@@ -36,7 +36,7 @@ def model() -> GameModel:
     return load_model(raw=build_game_data(json.loads(MODEL_FIXTURE.read_text())))
 
 
-def _scenario(**items: int) -> scn.Scenario:
+def _scenario(**items: float) -> scn.Scenario:
     return scn.from_dict({"name": "t", "items_produced": items, "rocket_launches": 1})
